+++ dependencies.gradle
@@ -4,8 +4,6 @@ dependencies {
     api("com.github.GTNewHorizons:GTNHLib:0.6.21:dev")
     implementation("com.github.GTNewHorizons:NotEnoughItems:2.7.40-GTNH:dev")
 
-    compileOnly("com.github.GTNewHorizons:ServerUtilities:2.1.46")
-
     testImplementation(platform('org.junit:junit-bom:5.9.2'))
     testImplementation('org.junit.jupiter:junit-jupiter')
