+++ repositories.gradle
@@ -1,4 +1,5 @@
 // Add any additional repositiroes for your dependencies here
 
 repositories {
+    mavenLocal()
 }
