+++ src/main/java/com/gtnewhorizon/structurelib/CommonProxy.java
@@ -67,8 +67,6 @@ public class CommonProxy {
 
     public void preInit(FMLPreInitializationEvent e) {}
 
-    public void init(FMLInitializationEvent e) {}
-
     public long getOverworldTime() {
         return MinecraftServer.getServer().getEntityWorld().getTotalWorldTime();
     }
